Add new options to server for supporting logging, simplify structure
